Bump distribute version to latest release
The older version of distribute failed on our Jenkins boxes, which are
running a fairly old version of pip, because they tried to patch the
installed version of pip and ended up introducing a reference to
an undefined variable.  (NameError: name 'install' is not defined)

This is fixed by upgrading to a newer version of distribute.