Re-enable nightly job
Re-enable the job which fetches page traffic from Google Analytics and
uses to update the popularity figures in the search indexes.

This was temporarily disabled because of a bug in the reindexing process
which replaced the Elasticsearch document `_type` values with the
default value. That bug has now been fixed, so this job is safe to
re-enable.