Remove now-irrelevant index cleanup step
The `rummager:clean` rake task deletes old indices, leaving only the
indices which currently have the alias. This step was necessary when the
popularity update job created a brand new index every night. The job now
updates the popularity values in place, so this deletion step is no
longer necessary.

Worse, the deletion step sometimes deletes old indexes that would be
useful to keep around for a day or two in case we want to switch back to
an old version of the index because there were problems in the overnight
job.

In the future, if someone creates a new index manually they will also
have to delete the old index manually (e.g. by running the rake task in
Jenkins).