Clean up page traffic index
We create a new page traffic index each night as part of the popularity update,
this used to be deleted as part of the nightly tasks, however we have disabled
the blanket delete so we don't remove backup indices.

This code will re-enable the delete for the page-traffic index only.