Fix one more str/bytes issue
This doesn't cause a problem when I run it locally, but does on
jenkins.  I've managed to run this part of the script on jenkins
directly, and have successfully fetched the traffic data.

I'm not sure why it's working locally though - different Python
version issue?  Jenkins uses 3.4, which is pretty old.