Update pip - add pylint and coverage - remove version on urllib3
We will be introducing tests; coverage and linter tools; and will recommend the use of venv (instead of virtualenv) as venv is part of the Python ‘batteries included’ ethos and ecosystem - this we will use where possible in order to reduce dependencies.

We are also adding pip as a named dependency.

During testing we found that the version given for urllib3 was causing dependency conflicts with other package versions. The simplest - and recommended - solution is to let the dependency system choose the most appropriate version. Doing this requires us to remove the version completely from this package.