Update README
We are updating the version of Python and recommending the use of pyenv. This is a clone of rbenv - but for Python. So, it should be familiar to all GDS developers and will help keep the local development environment as close as possible to the production one. Doing this should avoid the ‘it works on my machine’ problem.

We are also recommending the use of venv instead of virtualenv as per the previous commit. So, we update the documentation accordingly.

We will be introducing unit tests, coverage and linting tools. So we have chosen the tools that either: a) come bundled with Python - the Python ‘batteries included’ ethos; or are b) the most popular/frequently used by the community - in that order. We include documentation on how to run these tools, but assume developers will be familiar with these types of tool, the output they generate and the reasons and purpose of using them.