Adds unit tests for utils.py.
When writing the test for base64_encode, we found that passing in a bytes object resulted in a ‘TypeError: a bytes-like object is required, not 'str'’ error. This was due to the method trying to call .replace on a bytes object when it expected a string, despite the method parameter needing to be a bytes object.

The reason the method takes a bytes object is because it is used by AuthFileManager in auth.py for setting up the Google Analytics client. It writes env vars to a temp directory. This involves the env vars (strings) needing to be encoded and decoded. The env vars are compressed into a bytes object using zlib.compress and passed to base64_encode.

This code was last changed around 4 years ago when we were on Python 2.7. In Python 2 “the strings are by default treated as bytes” and “Python 3, by default, the strings are treated as Unicode. But unlike Python 2 there is no facility of automatic type coercion between Unicode strings and bytes” (https://www.stechies.com/typeerror-byteslike-object-required-not-str/#:~:text=This%20is%20a%20very%20common,on%20different%20versions%20of%20Python). Therefore we changed the method to replace bytes as opposed to strings within .replace.

It is not entirely clear why this wasn’t breaking in production. However, the job has been run locally with this change and the file is outputting as expected.