Restore dependencies to Feb 21st state
The bump of cryptography broke the process to fetch search analytics and
this was not discovered until nearly 3 months later. Given how risky it
is to update dependencies and how the useful life of this repo can be
measured in only weeks (due to Universal Analytics being sunset) it
seems reasonable to revert back.