Remove nightly-run.sh and update docs accordingly
nightly-run.sh has not been used since this project was updated to use
GitHub actions and is thus removed.

This modestly updates the readme to reflect the file is no longer
needed. There's still lots of TLC that could be done to this README to
reflect the wtfs in this project but where do you start?