Fix zlib.compress working on bytes
The zlib.compress function expects input to be bytes and outputs bytes.
Passing in a string returns an error of: TypeError: a bytes-like object is required, not 'str'

My expectation is that this is behaviour that changed in Python 3 and
that no-one has tried to run this code since using Python 2.