Use a local webserver for OAuth
Setting noauth_local_webserver to True avoids running a local webserver
and instead uses an out-of-band (OOB) OAuth flow. The problem with this is
that Google stopped supporting the OAuth OOB flow in January 2023.

These changes switch the creation of an OAuth token to be done with a
callback to a webserver on your machine running on localhost:8080 (as
per the defaults in the oauth2client module [1])

[1]: https://github.com/googleapis/oauth2client/blob/73d9d55447de97dfe541395817a0c8241701f7d6/oauth2client/tools.py#L61